refactor: consolidate training command logic and enhance configuration handling
- Streamlined the `train.py` command by delegating functionality to the `platform.cli.train` module, improving code organization and maintainability.
- Introduced a new `build_submit_summary_rows` function in `shared_config.py` to standardize summary row generation for both training and evaluation submissions.
- Updated the training configuration to utilize `AdvancedPassthroughSection` for advanced backend parameters, enhancing flexibility in configuration.
- Refactored the `submit` function to improve output path resolution and filename handling for training runs.
- Adjusted tests to ensure compatibility with the new structure and validate the updated command behavior.
