+++ tests/unit/cli/test_train_commands.py
@@ -107,7 +107,6 @@ def console_capture(monkeypatch: pytest.MonkeyPatch) -> StringIO:
     """
     output = StringIO()
     console = Console(file=output, force_terminal=False, width=200)
-    monkeypatch.setattr(train_module, "console", console)
     monkeypatch.setattr(platform_train_module, "console", console)
     monkeypatch.setattr(utils_module, "console", console)
     return output
@@ -137,6 +136,11 @@ def _mock_git_context(
 
     monkeypatch.setattr(
         "osmosis_ai.platform.cli.utils.require_git_workspace_directory_context",
+        _git_context,
+    )
+    monkeypatch.setattr(
+        platform_train_module,
+        "require_git_workspace_directory_context",
         _git_context,
     )
 
@@ -184,6 +188,7 @@ class TestListRuns:
                 )
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.list_runs(limit=30, all_=False)
 
         assert isinstance(result, ListResult)
@@ -214,6 +219,7 @@ class TestListRuns:
                 )
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.list_runs(limit=30, all_=False)
 
         assert isinstance(result, ListResult)
@@ -243,6 +249,7 @@ class TestListRuns:
                 )
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.list_runs(limit=30, all_=False)
 
         assert [column.label for column in result.columns] == [
@@ -289,6 +296,7 @@ class TestListRuns:
                 )
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.list_runs(limit=30, all_=False)
 
         assert isinstance(result, ListResult)
@@ -314,6 +322,7 @@ class TestListRuns:
                 )
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.list_runs(limit=30, all_=False)
 
         assert isinstance(result, ListResult)
@@ -399,6 +408,7 @@ class TestInfo:
                 return metric_data
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.info(name="run-1", output=str(tmp_path / "metrics.json"))
 
         assert isinstance(result, DetailResult)
@@ -437,6 +447,7 @@ class TestInfo:
                 raise AssertionError("pending info should not fetch metrics")
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.info(name="pending-run")
 
         assert isinstance(result, DetailResult)
@@ -480,6 +491,7 @@ class TestStatus:
                 raise PlatformAPIError("not available")
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.info(name="run-1")
 
         assert isinstance(result, DetailResult)
@@ -526,6 +538,7 @@ class TestStatus:
             get_training_run_metrics = _raise_metrics_unavailable
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.info(name="run-1")
 
         assert all(field.label != "Checkpoint" for field in result.fields)
@@ -573,6 +586,7 @@ class TestStatus:
             get_training_run_metrics = _raise_metrics_unavailable
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.info(name="run-1")
 
         assert result.sections
@@ -616,6 +630,7 @@ class TestStatus:
             get_training_run_metrics = _raise_metrics_unavailable
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.info(name="timed-run")
         fields = {field.label: field.value for field in result.fields}
 
@@ -653,6 +668,7 @@ class TestStatus:
             get_training_run_metrics = _raise_metrics_unavailable
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.info(name="full-run")
 
         assert isinstance(result, DetailResult)
@@ -724,6 +740,7 @@ rollout_batch_size = 64
                 return result
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         command_result = train_module.submit(config_path=config_path, yes=True)
 
         assert isinstance(command_result, OperationResult)
@@ -761,6 +778,7 @@ rollout_batch_size = 64
                 return result
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         command_result = train_module.submit(config_path=config_path, yes=True)
 
         expected_url = result.platform_url
@@ -788,6 +806,7 @@ rollout_batch_size = 64
 
         FakeClient.SUBMIT_RESULT = self.SUBMIT_RESULT
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.submit(config_path=config_path, yes=True)
         out = console_capture.getvalue()
         assert "calculator" in out
@@ -837,6 +856,7 @@ rollout_batch_size = 64
                 return TestSubmit.SUBMIT_RESULT
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.submit(config_path=path, yes=True)
         assert captured_kwargs["experiment_config"]["commit_sha"] == "deadbeef"
         assert captured_kwargs["git_identity"] == GIT_IDENTITY
@@ -862,6 +882,7 @@ rollout_batch_size = 64
                 return TestSubmit.SUBMIT_RESULT
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         train_module.submit(config_path=config_path, yes=True)
         assert captured_kwargs["experiment_config"] == {
             "model_path": "Qwen/Qwen3.6-35B-A3B",
@@ -914,6 +935,7 @@ OPENAI_API_KEY = "openai-api-key"
                 return TestSubmit.SUBMIT_RESULT
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         train_module.submit(config_path=path, yes=True)
 
         assert captured_kwargs["env_config"] == {"LOG_LEVEL": "INFO"}
@@ -996,6 +1018,7 @@ rollout_batch_size = 64
                 return TestSubmit.SUBMIT_RESULT
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         train_module.submit(config_path=config_path, yes=True)
 
         out = console_capture.getvalue()
@@ -1033,6 +1056,7 @@ rollout_batch_size = 64
                 return TestSubmit.SUBMIT_RESULT
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         train_module.submit(config_path=config_path, yes=True)
 
         out = console_capture.getvalue()
@@ -1069,6 +1093,7 @@ rollout_batch_size = 64
                 return TestSubmit.SUBMIT_RESULT
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         train_module.submit(config_path=config_path, yes=True)
 
         out = console_capture.getvalue()
@@ -1105,6 +1130,7 @@ rollout_batch_size = 64
                 return TestSubmit.SUBMIT_RESULT
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         train_module.submit(config_path=config_path, yes=True)
 
         out = console_capture.getvalue()
@@ -1144,6 +1170,7 @@ rollout_batch_size = 64
                 return TestSubmit.SUBMIT_RESULT
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         train_module.submit(config_path=path, yes=True)
 
         out = console_capture.getvalue()
@@ -1166,6 +1193,7 @@ rollout_batch_size = 64
                 return TestSubmit.SUBMIT_RESULT
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.submit(config_path=config_path, yes=True)
 
         assert isinstance(result, OperationResult)
@@ -1261,6 +1289,7 @@ def test_train_submit_accepts_project_subdirectory(
             return TestSubmit.SUBMIT_RESULT
 
     monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+    monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
     rc = main(["--json", "train", "submit", "configs/training/default.toml", "--yes"])
     captured = capsys.readouterr()
 
@@ -1291,6 +1320,7 @@ class TestSubmitNonInteractiveContext:
                 raise AssertionError("API must not be reached without confirmation")
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
 
         from osmosis_ai.cli import main as cli
 
@@ -1363,6 +1393,7 @@ class TestSubmitNonInteractiveContext:
                 raise AssertionError("API must not be reached without confirmation")
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
 
         from osmosis_ai.cli import main as cli
 
@@ -1444,6 +1475,7 @@ class TestMetrics:
                 return metrics
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.info(
             name="run-1",
             output=str(tmp_path / "metrics.json"),
@@ -1472,6 +1504,7 @@ class TestStop:
                 return {"stopped": True}
 
         monkeypatch.setattr(api_client_module, "OsmosisClient", FakeClient)
+        monkeypatch.setattr(platform_train_module, "OsmosisClient", FakeClient)
         result = train_module.stop(name="my-run", yes=True)
 
         assert isinstance(result, OperationResult)
